backend/requirements.txt
flask
flask-cors
torch
segment-anything
numpy
opencv-python
Pillow
